+++ core/js/typespecs/functions.ts
@@ -6,7 +6,6 @@ import { chatCompletionMessageParamSchema } from "./openai/messages";
 import { customTypes } from "./custom_types";
 import { gitMetadataSettingsSchema, repoInfoSchema } from "./git_types";
 import { objectReferenceSchema } from "./common_types";
-import { functionDataSchema } from "./app_types";
 
 export const validRuntimesEnum = z.enum(["node", "python"]);
 export type Runtime = z.infer<typeof validRuntimesEnum>;
@@ -26,6 +25,75 @@ const strictParam = z
   .describe(
     "If true, throw an error if one of the variables in the prompt is not present in the input",
   );
+export const codeBundleSchema = z
+  .object({
+    runtime_context: runtimeContextSchema,
+    location: z.union([
+      z
+        .object({
+          type: z.literal("experiment"),
+          eval_name: z.string(),
+          position: z.union([
+            z.object({ type: z.literal("task") }),
+            z
+              .object({
+                type: z.literal("scorer"),
+                index: z.number().int().nonnegative(),
+              })
+              .openapi({ title: "scorer" }),
+          ]),
+        })
+        .openapi({ title: "experiment" }),
+      z
+        .object({
+          type: z.literal("function"),
+          index: z.number().int().nonnegative(),
+        })
+        .openapi({ title: "function" }),
+    ]),
+    bundle_id: z.string(),
+    preview: z.string().nullish().describe("A preview of the code"),
+  })
+  .openapi("CodeBundle");
+export type CodeBundle = z.infer<typeof codeBundleSchema>;
+
+export const functionDataSchema = z
+  .union([
+    z
+      .object({
+        type: z.literal("prompt"),
+        // For backwards compatibility reasons, the prompt definition is hoisted out and stored
+        // in the outer object
+      })
+      .openapi({ title: "prompt" }),
+    z
+      .object({
+        type: z.literal("code"),
+        data: z.union([
+          z
+            .object({
+              type: z.literal("bundle"),
+            })
+            .and(codeBundleSchema)
+            .openapi({ title: "bundle" }),
+          z
+            .object({
+              type: z.literal("inline"),
+              runtime_context: runtimeContextSchema,
+              code: z.string(),
+            })
+            .openapi({ title: "inline" }),
+        ]),
+      })
+      .openapi({ title: "code" }),
+    z
+      .object({
+        type: z.literal("global"),
+        name: z.string(),
+      })
+      .openapi({ title: "global" }),
+  ])
+  .openapi("FunctionData");
 
 export const functionIdSchema = z
   .union([
